+++ test/Log/AccessLogFactoryTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-swoole for the canonical source repository
- * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2018-2019 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-swoole/blob/master/LICENSE.md New BSD License
  */
 
@@ -10,8 +10,6 @@ declare(strict_types=1);
 namespace ZendTest\Expressive\Swoole\Log;
 
 use PHPUnit\Framework\TestCase;
-use Psr\Container\ContainerInterface;
-use Psr\Log\LoggerInterface;
 use ReflectionProperty;
 use Zend\Expressive\Swoole\Log\AccessLogFactory;
 use Zend\Expressive\Swoole\Log\AccessLogFormatter;
@@ -21,25 +19,16 @@ use Zend\Expressive\Swoole\Log\StdoutLogger;
 
 class AccessLogFactoryTest extends TestCase
 {
-    public function setUp()
-    {
-        $this->container = $this->prophesize(ContainerInterface::class);
-        $this->logger = $this->prophesize(LoggerInterface::class)->reveal();
-        $this->formatter = $this->prophesize(AccessLogFormatterInterface::class)->reveal();
-    }
+    use LoggerFactoryHelperTrait;
 
     public function testCreatesDecoratorWithStdoutLoggerAndAccessLogFormatterWhenNoConfigLoggerOrFormatterPresent()
     {
         $factory = new AccessLogFactory();
 
-        $this->container->has('config')->willReturn(false);
-        $this->container->get('config')->shouldNotBeCalled();
-        $this->container->has(LoggerInterface::class)->willReturn(false);
-        $this->container->get(LoggerInterface::class)->shouldNotBeCalled();
         $this->container->has(AccessLogFormatterInterface::class)->willReturn(false);
         $this->container->get(AccessLogFormatterInterface::class)->shouldNotBeCalled();
 
-        $logger = $factory($this->container->reveal());
+        $logger = $factory($this->createContainerMockWithConfigAndNotPsrLogger());
 
         $this->assertInstanceOf(Psr3AccessLogDecorator::class, $logger);
         $this->assertAttributeInstanceOf(StdoutLogger::class, 'logger', $logger);
@@ -50,14 +39,10 @@ class AccessLogFactoryTest extends TestCase
     {
         $factory = new AccessLogFactory();
 
-        $this->container->has('config')->willReturn(false);
-        $this->container->get('config')->shouldNotBeCalled();
-        $this->container->has(LoggerInterface::class)->willReturn(true);
-        $this->container->get(LoggerInterface::class)->willReturn($this->logger);
         $this->container->has(AccessLogFormatterInterface::class)->willReturn(false);
         $this->container->get(AccessLogFormatterInterface::class)->shouldNotBeCalled();
 
-        $logger = $factory($this->container->reveal());
+        $logger = $factory($this->createContainerMockWithConfigAndPsrLogger());
 
         $this->assertInstanceOf(Psr3AccessLogDecorator::class, $logger);
         $this->assertAttributeSame($this->logger, 'logger', $logger);
@@ -68,21 +53,10 @@ class AccessLogFactoryTest extends TestCase
     {
         $factory = new AccessLogFactory();
 
-        $this->container->has('config')->willReturn(true);
-        $this->container->get('config')->willReturn([
-            'zend-expressive-swoole' => [
-                'swoole-http-server' => [
-                    'logger' => [
-                        'logger-name' => 'my_logger',
-                    ],
-                ],
-            ],
-        ]);
-        $this->container->get('my_logger')->willReturn($this->logger);
         $this->container->has(AccessLogFormatterInterface::class)->willReturn(false);
         $this->container->get(AccessLogFormatterInterface::class)->shouldNotBeCalled();
 
-        $logger = $factory($this->container->reveal());
+        $logger = $factory($this->createContainerMockWithNamedLogger());
 
         $this->assertInstanceOf(Psr3AccessLogDecorator::class, $logger);
         $this->assertAttributeSame($this->logger, 'logger', $logger);
@@ -93,14 +67,10 @@ class AccessLogFactoryTest extends TestCase
     {
         $factory = new AccessLogFactory();
 
-        $this->container->has('config')->willReturn(false);
-        $this->container->get('config')->shouldNotBeCalled();
-        $this->container->has(LoggerInterface::class)->willReturn(false);
-        $this->container->get(LoggerInterface::class)->shouldNotBeCalled();
         $this->container->has(AccessLogFormatterInterface::class)->willReturn(true);
         $this->container->get(AccessLogFormatterInterface::class)->willReturn($this->formatter);
 
-        $logger = $factory($this->container->reveal());
+        $logger = $factory($this->createContainerMockWithConfigAndNotPsrLogger());
 
         $this->assertInstanceOf(Psr3AccessLogDecorator::class, $logger);
         $this->assertAttributeInstanceOf(StdoutLogger::class, 'logger', $logger);
@@ -111,8 +81,10 @@ class AccessLogFactoryTest extends TestCase
     {
         $factory = new AccessLogFactory();
 
-        $this->container->has('config')->willReturn(true);
-        $this->container->get('config')->willReturn([
+        $this->container->has(AccessLogFormatterInterface::class)->willReturn(false);
+        $this->container->get(AccessLogFormatterInterface::class)->shouldNotBeCalled();
+
+        $logger = $factory($this->createContainerMockWithConfigAndNotPsrLogger([
             'zend-expressive-swoole' => [
                 'swoole-http-server' => [
                     'logger' => [
@@ -121,13 +93,7 @@ class AccessLogFactoryTest extends TestCase
                     ],
                 ],
             ],
-        ]);
-        $this->container->has(LoggerInterface::class)->willReturn(false);
-        $this->container->get(LoggerInterface::class)->shouldNotBeCalled();
-        $this->container->has(AccessLogFormatterInterface::class)->willReturn(false);
-        $this->container->get(AccessLogFormatterInterface::class)->shouldNotBeCalled();
-
-        $logger = $factory($this->container->reveal());
+        ]));
 
         $this->assertInstanceOf(Psr3AccessLogDecorator::class, $logger);
         $this->assertAttributeInstanceOf(StdoutLogger::class, 'logger', $logger);